eventgrid-backend/requirements.txt
flask==3.0.3
flask-cors==5.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1

